# server/src/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Games API",
    "version": "1.0.0",
    "description": "This API provides endpoints for managing games in a database. The /games endpoint retrieves a list of all games, while the /games/{id} endpoint retrieves a specific game by ID. The /game endpoint allows for searching games by name, while /api/games-top retrieves the top 10 games by recommendation percentage. The /api/games endpoint creates a new game in the database, while the /api/games/{id} endpoint updates an existing game. The API is built using Swagger and follows the OpenAPI specification. The Game schema defines the structure of a game object, including properties such as ID, recommendation percentage, number of reviews, top critic score, tier, name, first release date, and URL. The API returns responses in JSON format and provides error messages for invalid requests."
  },
  "tags": [
    {
      "name": "Games",
      "description": "API for games in the system"
    }
  ],

  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "components": {
    "schemas": {
      "Game": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "percentRecommended": {
            "type": "number",
            "format": "float"
          },
          "numReviews": {
            "type": "integer",
            "format": "int32"
          },
          "topCriticScore": {
            "type": "number",
            "format": "float"
          },
          "tier": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "firstReleaseDate": {
            "type": "integer",
            "format": "int32"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "percentRecommended",
          "numReviews",
          "topCriticScore",
          "tier",
          "name",
          "firstReleaseDate",
          "url"
        ]
      }
    }
  }
,
  "paths": {
    "/api/all-games": {
      "get": {
        "summary": "Retrieve all games.",
        "description": "Get all games from database",
        "tags": ["Games"],
        "responses": {
          "200": {
            "description": "A list of games.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Game"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Error message."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },

    "/api/game/{id}": {
      "get": {
        "summary": "Get a game by ID",
        "description": "Get a game from the database by its ID",
        "tags": ["Games"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the game to retrieve",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The game object that matches the given ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Game"
                }
              }
            }
          },
          "500": {
            "description": "Game not found"
           
          }
        }
      }
    },
    "/api/games": {
      "get": {
        "summary": "Get games by name",
        "description": "Get a list of games by name from the database",
        "tags": ["Games"],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            },
            "description": "The name of the game to search for",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "A list of games that match the search criteria",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Game"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Game not found"
           
          }
          
        }
      }
    },
    "/api/games-top": {
      "get": {
        "summary": "Get top 10 games by recommendation percentage.",
        "tags": ["Games"],
        "responses": {
          "200": {
            "description": "Returns an array of top 10 games.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Game"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Games not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/add-game": {
      "post": {
        "summary": "Create a new game",
        "tags": ["Games"],
        "requestBody": {
          "description": "Object containing the game data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Game"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully created a new game",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Game"
                }
              }
            }
          },
          "500": {
            "description": "Failed to create game",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/update-game/{id}": {
      "put": {
        "summary": "Update a game by ID",
        "description": "Update a game in the database using its ID",
        "tags": ["Games"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the game to update",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Updated game object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Game"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK. Returns the updated game object.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Game"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. Invalid input parameters."
          },
          "404": {
            "description": "Not Found. Game not found."
          },
          "500": {
            "description": "Internal Server Error. Failed to update game in the database.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Game": {
      "type": "object",
      "required": [
       
        "percentRecommended",
        "numReviews",
        "topCriticScore",
        "tier",
        "name",
        "firstReleaseDate",
        "url"
      ],
      "properties": {
        "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique identifier for the game",
            "example": 1
          },
          "percentRecommended": {
            "type": "number",
            "format": "float",
            "example": 60.4
          },
          "numReviews": {
            "type": "integer",
            "example": 500
          },
          "topCriticScore": {
            "type": "number",
            "format": "float",
            "example": 67
          },
          "tier": {
            "type": "string",
            "example": "Mighty"
          },
          "name": {
            "type": "string",
            "example": "Grand Theft Auto V"
          },
          "firstReleaseDate": {
            "type": "integer",
            "example": 2012
          },
          "url": {
            "type": "string",
            "example": "https://www.rockstargames.com/V/"
          }
      }
    }
  }
}